style: rename components.
